+++ index.js
@@ -1,4 +1,4 @@
-exports.version = '2.2.0';
+exports.version = '2.3.0';
 
 exports.cache = {};
 exports.user = null